requirements.txt
# ============================================================
#  RetinaGuard — Python Dependencies
#  Install with:  pip install -r requirements.txt
# ============================================================

# ── Web Framework ─────────────────────────────────────────────
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.4

# ── Screening Engine ─────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Image Processing ──────────────────────────────────────────
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# ── Data Utilities ────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ── Dev Tools (optional, remove for production) ───────────────
# notebook
# ipywidgets
